nix: use bazelisk without influence from CC and CXX envvars
This is the result of William and I debugging getting bazel to work on
my nix-darwin managed M2 macbook to work with bazel. This is the output
of a full morning of work :)

The root cause of the issue we ran into was we could not get bazel to
compile protobuf, it would fail to find the C++ std headers. We had many
misteps here, it turned out that unsetting CC and CXX would lead to
bazel using the system installed toolchain. We are unsure why we can't
use nix's clang-wrapped, maybe it has to do with that respecting special
NIX_ envvars for building nix packages?

Test Plan: "bazel test //cmd/frontend/internal/search:search_test" and
"bazel configure" works.