extsvc: always retry if receiving a StatusTooManyRequests
This is the likely cause of a flake we saw in CI. It is in fact a real
issue that can happen when the retry-at timestamp we get back is close
to our current time. This leads to a race were the timestamp we get back
from Azure or GitLab is in the past according to our clock. In the
monitor API this would lead us to not retrying even though we can.

We attempted to improve the API of monitor to allow zero sleeps. But
this likely breaks GitHub were we don't always get back rate limit
headers. As such we updated the callsites in Azure DevOps and GitLab.

Test Plan: go test

Co-authored-by: Petri-Johan Last <petri.last@sourcegraph.com>