searcher: always call Wait on comby processes (#50034)
If you call Start you _must_ always call Wait to free up resources. In
particular this is the likely cause of a huge goroutine leak in the
searcher process which all blames to the os/exec package waiting on a
context.

This was an accidental regression introduced when this code was ported
to using conc. Previously it would always call Wait, but now if
something in the introduced pool fails, we don't call Wait.

Test Plan: CI

(cherry picked from commit 257ba04b6498785584915fcec435244b3ee641c7)