fix: grant pending permissions not working correctly (#50059)
## Description

GrantPendingPermissions can be called multiple times in a loop. When
using unified perms, it uses a method that by default removed old
records from the table, effectively replacing existing records with new
records.

However this behavior is not correct for granting pending permissions,
which should only add new records, not remove old ones. Fixing this
allows us to call GrantPendingPermissions in a loop without causing
incosistencies in the database.

## Test plan

extensively unit tested so that we are sure it works for both legacy and
unified permissions