file view: Fix in-file search behavior wrt selections (#49968)
There are a couple of issues with our custom in-file search UI
implementation:

- Selecting a word and then triggering a search will highlight the
matches but not populate the UI correctly. The input is empty and no
result count is shown.
- Similarly if the search is already open (and shows correct data),
selecting a different term in the file and triggering the search will
update the match highlights but not the input or result count.
- Navigating to the next/previous match via the keyboard (e.g. with
`F3`) does not update the selected match index.


I believe the core problems with the current implementation are:

- The input is not controlled, i.e. we are not deriving the value from
`searchQuery`. This causes the input not to be populated correctly when
the search query is changed externally.
- The matches and selected match index are updated from state fields,
which in turn are updated when the input changes. This is problematic
for two reasons:
- As in the previous point, the UI will show incorrect values when the
search query is changed externally.
- Even when we compute the matches and index when the component is first
created, the (outdated) state field values will overwrite the correct
values on the next update.

This PR makes the following changes:

- Remove state fields for matches and selected index in favor of
computing those when the searchQuery changes. I don't think state fields
are necessary here, the view component can track this state itself. It
also makes data flow simpler.
- Update the selected index when the selection changes. This also means
that the index is cleared when the user selects a different text while
the search is open. Given that CodeMirror changes the highlighting
though I'd say this is the more correct behavior.
- The UI will now show `x / x results` instead of just `x / x`. This is
the consequence of simplifying how this part is rendered. I can change
that if it's not desired.

## Test plan

- Open a file, select a word and press `Mod-F` -> search component opens
with populated input and correct result count
- Click anywhere outside of a highlight -> result count updates from `x
/ x results` to `x results`.
- Click previous/next buttons -> result count updates to `x / x results`
- Press F3 to go to the next match -> selected result updates