scip-typescript: upload index to s2
Previously, we didn't upload scip-typescript indexes to S2 resulting in
different navigation on S2 and dotcom. This PR fixes the immediate issue
and in this quarter we are planning to fix the auto-indexing inference
script so that we no longer need to maintain a custom scip-typescript CI
job for the sourcegraph/sourcegraph monorepo.