[feat][force-bazel] make unified perms not experimental (#49971)
## Description

Make unified perms the only way to read permissions. This PR cleans up
the legacy way of doing things in most places, but for the sake of
trying to keep it shorter, only tackles the read path. So it still keeps
some of the legacy methods in perms_syncer.go and perms_store.go. I want
to create a followup PR for cleaning that up.

## Test plan

Unit tests, which already cover all the changed functionality.