scip-typescript: upload index to S2 (#50077)
Previously, we didn't upload scip-typescript indexes to S2 resulting in
different navigation on S2 and dotcom. This PR fixes the immediate issue
and in this quarter we are planning to fix the auto-indexing inference
script so that we no longer need to maintain a custom scip-typescript CI
job for the sourcegraph/sourcegraph monorepo.


## Test plan
Verify that the scip-typescript CI job uploads to S2.

<!-- All pull requests REQUIRE a test plan:
https://docs.sourcegraph.com/dev/background-information/testing_principles
-->

## App preview:

- [Web](https://sg-web-olafurpg-upload-s2.onrender.com/search)

Check out the [client app preview
documentation](https://docs.sourcegraph.com/dev/how-to/client_pr_previews)
to learn more.