cody: add debug logger
This adds a singleton logger that allows us to easily log to the debug
tab without much overhead in production mode.

Why can we not use vscodeAPI directly?
The logger needs access to acquireVsCodeAPI which is not available
codebase-context/index.ts. Hence we set the logging func in
ChatViewProvider which has access to it.

As a first example, we log the context type and the context
metadata (see client/cody-shared/src/codebase-context/index.ts)

Test plan:
- Set debug=true in cody's vscode settings, reload the window
- Verify that the debug window appears
- Ask cody a question
- Check the debug tab for output