fix unified perms oob migrator (#50147)
Turns out our progress was not reporting correctly, because certain rows
were not migrated. This was caused by bad query in migrator itself,
because if there were no permissions (empty `object_ids_ints` array), we
had no rows in the `s` set of the subquery. This `s` subquery was
subsequently used to update the `migrated` column, but we should have
instead used the `candidates` subquery.

Hope the description makes sense, code change should be self
explanatory.

## Test plan

- [x] Unit tested
- [x] Verify on S2 that this is the case