embeddings: per repo mutex when getting/downloading an index (#50241)
Acquire a mutex for the given repository to serialize access when
downloading/getting the repo embedding index. This avoids multiple
routines concurrently downloading the same embedding index when the
cache is empty, which can lead to an out-of-memory error.

## Test plan

* Tested with a script creating multiple concurrent requests. (Found a
deadlock in the original commit that way and fixed it.)

(cherry picked from commit fb3615e170add2b72eb22e6b8f4d9a1dad60000e)