extract Cody UI package, add Cody to web app (#50270)
- The new `client/cody-ui` package (`@sourcegraph/cody-ui`) contains
reusable components for Cody's UI that are shared among all web UIs for
Cody.
- The Sourcegraph web app repo sidebar now adds a `Cody` tab (alongside
files and symbols) when the `cody-experimental` feature flag is enabled.
This lets you use Cody in the web app.
- A new demonstration `client/cody-web` package implements a standalone
web app for Cody.

No behavior change is intended to the Cody code.

## Test plan

Build and run the Cody extension locally. Ensure that it still functions
as is. All other changes are experimental and behind a feature flag (in
the main web app).