cmd/gitserver: Remove scrubRemoteURL cleanup job (#50272)
This was added in https://github.com/sourcegraph/sourcegraph/pull/18535
to clean up remote URL that was being set on disk. A long time since
then has passed and we don't set remote on disk any longer. This makes
it a safe change to remove for both existing Sourcegraph instances (all
existing repos would have been cleaned out since release v3.26 was
rolled out) and for new Sourcegraph that were created after this change
was implemented, this is a NOOP.