[Backport 5.0] Code monitors: truncate long Slack blocks (#50309)
Slack requires that text blocks be no longer than 3000 characters. We
already limit the number of lines in the match blocks, but if the lines
are long then the block can still exceed the character limit. With this
change, we also truncate lines if they will push the block over the
character limit.

Fixes #49979.

## Test plan

New unit test for character-based truncation.

Also tested manually -- reproduced the error, then verified this change
resolves it.