build frontend based on bzl artifact  (#49924)
Introduce a conditional path to build both frontends using Bazel
produced artifacts instead of the normal ones.

This is achieved by introducing a build setting named
`//:assets_bundle_type` which can take either `oss` or `enterprise`.
When enabled, this settings flips which files are included in the
filegroup that is embedded in the `//ui/assets:assets` go library. It
also automatically add the `dist` go tags to the frontend binaries,
which in turns pick the right sources in the `ui/assets` Go package.

If it's not set, it defaults to empty, which will instead run the dev
variant of the assets, expecting the files to be maintained on the FS
during the runtime, usually by `bundle-dev`.

This has a side effect of invalidating the analysis cache, but it's
fairly acceptable at this point. In the future, we want to refactor the
assets handling code to instead have two explicit dependencies whenever
that oss and enterprise pick accordingly, making the dependency graph
static regardless of the bundle type.

## Test plan

<!-- All pull requests REQUIRE a test plan:
https://docs.sourcegraph.com/dev/background-information/testing_principles
-->

This was tested by running `sg start oss-bazel` and manually specifying
the right `--//:assets_bundle_type=X` to ensure that we get the correct
assets and that it stil works without it, in a classic dev setting.