embeddings: disable document ranks per default (#50296)
This PRs adds the boolean 'useDocumentRanks=FALSE' to the API of
embeddings search.

We have seen confusing results for context fetching on s2 that can be
explained by missing or broken document ranks. The background job to
calculate path ranks is still in beta. Hence, for now, we disable the
contribution of document ranks to the score.