Go to definition: push current location into browser history
Previously, the behavior was different whether you clicked on the "Go to definition"
link in the popover or used cmd-click. When using the link, the current
location didn't get pushed into the browser history making "Go back"
jump to a surprising location. When using cmd-click to navigate to a
definition, we pushed the current location into the browser history to
replicate the experience of navigating code in the editor where you can
easily go up/down the "jump stack", which is the list of locations you
navigate from/to with "Go to definition".

Importantly, this PR preserves the HTML link in the popover so that
users can still right-click "Open link in new tab".