Display cloning progress (#50234)
Part of #39327

<img width="1194" alt="Screenshot 2023-04-03 at 8 43 03 PM"
src="https://user-images.githubusercontent.com/153410/229665035-3572524e-1473-482c-970f-45d56c13f098.png">

Repositories under site admin display cloning progress in the bottom
line while cloning. The data is only queried when
`clone-progress-logging` feature flag is on. This is important, as when
the feature flag is off, the request for cloning progress goes straight
to gitserver. With the feature flag on the back-end goes to the database
to retrieve the data.

## Test plan

UI: Visual inspection.

Backend: Unit testing.

Loom also examining the feature flag:

https://www.loom.com/share/440c0fd88f6646079762a8dcfb4bdebc