dep: Update Grafana alpine base image (#50376)
The Grafana Alpine docker base image did not get updated last time. This
is due to the naming of the Dockerfile. I've manually updated the
base-image and updated the glob in the `update-all-dockerfiles.sh`
script.

## Test plan
Ran a build locally, the CVE does not occur in this image. Tested the
glob, by running the script and confirmed that the alpine Dockerfile was
included.

<!-- All pull requests REQUIRE a test plan:
https://docs.sourcegraph.com/dev/background-information/testing_principles
-->