ci: skip bazel build //... before bazel test //... (#50436)
It is safe to only run `bazel test //...` in CI instead of `bazel build
//...` followed by `bazel test //...`. Bazel will implicitly build any
targets needed for testing as part of the bazel test command. Running
the explicit bazel build //... first does not provide any benefits and
only adds build time to the CI cycle.

Bazel builds have two phases:
1. The analysis phase, where Bazel determines which targets need to be
built/rebuilt based on changes.
2. The execution phase, where the actual building and testing happens.
When you run `bazel test //...`, Bazel will do a full analysis of the
codebase and determine which tests need to be built and run based on
what's changed. It will then build any dependencies of those tests on
the fly during the execution phase, as needed to actually run the tests.
So a full `bazel build //...` is not needed beforehand - Bazel is smart
enough to figure out what needs to be built for the testing to be done.
By only running `bazel test //...`, we save the time of rebuilding the
entire codebase, and only build what's needed to run the tests that are
affected by changes. Therefore, we simplify our CI builds to only a
single `bazel test //...` step.

This will make our CI builds faster and simpler, while still ensuring
thorough test coverage of changes.



## Test plan

<!-- All pull requests REQUIRE a test plan:
https://docs.sourcegraph.com/dev/background-information/testing_principles
-->

CI