convert all errors
We convert RPC errors into well-typed errors at the API boundary, but we
only did this for errors returned from Recv before. This adds the
conversion to the Search call. We switch on the error type higher up in
the job tree, so it's necessary to convert these errors to proper
context errors.