Cody: Add a recipe to "replace" the selection.
This command is a new way of working with Cody that is half way
between chat and autocomplete. It is a bit like pair programming or
code reviewing. You write instructions in place in source text, for
example:

- "we need error checking here"
- "move this state inside the iterator and make it IterMut"
- "this fails with an error that 'shader program is not linked'"

Then select the area of the program to change and invoke "Ask Cody:
Replace Instructions in Selection." Cody will act on the selection. As
with any Cody dialogue, if Cody gets wedged or confused you can start
a fresh conversation and try again.

This adds a BotResponseMultiplexer where recipes can name topics and
subscribe to the agent's incremental output in these topics. This lets
recipes act on the agent's response while the agent can also stream
commentary into chat.

Fixes #50281

 ## Test plan

```sh
$ cd client/cody-shared
$ pnpm test
$ cd ../cody
$ pnpm test:unit
$ pnpm test:integration
```

Plus manually test the command. Automated testing is blocked on #50118