remove code host integration messaging setting
This was used to configure whether users should be informed of the browser extension or the native integration. However, it only affected the bottom-most nav item in the user menu, which is not very prominent. (Previously, it showed up in more prominent places, but those notices were annoying and were removed.)

Because this only affected a low-prominence thing in the UI, it is not worth keeping. Removing this also simplifies the user menu.