SCIM: Add support for soft delete (#50533)
When modifying users using PUT/PATCH via SCIM toggling the `active`
property should have the effect of soft deleting and restoring the user.

This code first updates the SCIM logic to remove any user specific
entity logic from the generic logic that would apply to any scim entity
such as how to manipulate attributes. The effect of this is that each
possible type of change is separated, and runs a check on the attributes
before/after the changes to determine if it has occurred and an action
needs to be taken. After this change I replaced the existing hard delete
code for `active` changes with a soft delete/reactivate.