Cody: fix bug where currently open file dominated context (#50572)
The change in #50101 caused a regression where Cody would refuse to
answer questions about the codebase when there was a file open in the
editor. This was because Cody thought the open file was the only context
it was allowed to use. This PR adjusts the editor prompt so Cody will
use all context files.