Add python to bundled executor (#50617)
Adding this on customer request. Trivy seems not concerned about this,
so this should be fine. It's a one-off solution still so doing customer
specific things here should be alright. xmlstarlet was another of the
things we have added here.

## Test plan

Verified trivy, built the image locally and checked that python works.