remove code host integration messaging setting (#50582)
This was used to configure whether users should be informed of the
browser extension or the native integration. However, it only affected
the bottom-most nav item in the user menu, which is not very prominent.
(Previously, it showed up in more prominent places, but those notices
were annoying and were removed.)

Because this only affected a low-prominence thing in the UI, it is not
worth keeping. Removing this also simplifies the user menu.



![image](https://user-images.githubusercontent.com/1976/231702997-32cc4b45-9c73-406f-84ab-15f546c784a7.png)


## Test plan

Automated tests will catch issues.