show action chips for recipes instead of just free-form text
For example, instead of a (false summary) `Replace the instructions in the selection`, Cody now shows a chip `Fixup (filename)` that represents the action the user took (running the fixup recipe).

When Cody responds, instead of a vague sentence, it can show a chip to indicate what it did (but this is not yet implemented).