Backport 50668 to 5.0 (#50720)
This backports both 
- #49920
- #50668 

to 5.0. The former is cherry picked to reduce merge conflicts and make
potential future cherry picks a bit easier.
I also had to backport adding a bazelignore entry for `client/cody`
because I couldn't update the build files on this branch anymore. There
seems to have been a few misses in buildfile updates so this PR is
unfortunately quite noisy.

## Test plan

CI and code review.

---------

Co-authored-by: Rok Novosel <rok@sourcegraph.com>
Co-authored-by: Beyang Liu <beyang@sourcegraph.com>