embeddings: filter out missing files (#50810)
Currently if we fail to find a file we do not remove it from the
results. Instead we keep the result without Contents populated. This has
the potential to leak filenames in the case of sub-repo permissions.
However, I haven't validated if those are enforced in another layer /
are even enforced by readFileFn. As such this is just a more correct
implementation / less surprising return results.

Additionally if StartLine or EndLine do get clamped, we update the
result to reflect this.

We pull this logic out into its own function to make it clear how
filtering is done. Additionally we intend to call this function for the
experimental weaviate support.

Test Plan: go test