Implement rate limiting for completions API
For dotcom, we want to enforce per-user rate limits. This rate limiter implementation will also be mostly usable in the LLM proxy code. When configured, a user will have to stick to the rate limit. Unauthenticated users will be rate limited by IP address. I've added the ability to increase or decrease the rate limit for an individual user.