Allow site admins on Sourcegraph dotcom to view profiles
This PR changes NOTHING about access permissions. All resolvers still respect the same permissions applied as before, so what was not possible before is still not possible.
What this PR does is unhide a bunch of UI components that actually were allowed but hidden by this not very permissive check.
Before, the user settings pages were completely hidden (albeit accessible through the API easily). Now, the UIs that are permitted by the backend permission checks are visible. That includes the list of access tokens, management features for executor secrets, and such. This can be a useful tool for debugging, and will allow us to manage other user properties that ARE meant to be configurable by the site admin.