convert to int8 and unroll loop
This converts our embeddings to an int8 in memory. It does not convert
the storage format so that we can be backwards compatible with the
embeddings stored in blob storage. Also, I could see it being useful to
maintain the more precise embeddings in case we see regressions in
ranking from this change.