show "reset conversation" icon in view/title menu not react UI
The standard place for this kind of button is in the view/title menu, not in the webview UI. This change also makes it use the standard VS Code icon for this type of action, instead of a non-VS Code-standard icon (the reload).

This is similar to how we moved settings to a gear icon in the sme view/title menu.