completions: factor out observability stuff for resolver
Third time is the charm. Instead of more copy-pasta, I factored out the
pattern. Additionally I made them all log to the same honeycomb dataset
since now that I have learnt more it will be useful to have aggregate
information here.

The reason we go with a builder pattern is because it makes it nice to
optionally add stuff. For example it is used here since we don't have
the http request in the graphql resolvers.

Test Plan: go test