Implement rate limiting for completions API (#50925)
For dotcom, we want to enforce per-user rate limits. This rate limiter
implementation will also be mostly usable in the LLM proxy code. When
configured, a user will have to stick to the rate limit. Unauthenticated
users will be rate limited by IP address. I've added the ability to
increase or decrease the rate limit for an individual user.

Manually verified things work as expected and added some tests. I have
not added tests for the request paths as this code is all untested and I
won't be able to get to it E2E before PTO, if we feel comfortable with
the level of manual testing I've done here, we can merge this and get it
out.

Co-authored-by: Keegan Carruthers-Smith <keegan.csmith@gmail.com>
(cherry picked from commit d618b61edf10dfcb50bce4e234f15d731cd226ef)