[app] Use single, embedded DB for development
This was a weird/surprising issue. The problem was that I wasn't able to
run `sg start app` without having an external DB running. This was
surprising especially since #48020 should have enabled that.
Contrary to what I said in #50859, the embedded DB _is_ is used, but
only for the frontend. CodeIntel and CodeInsights require an external
DB.
Why? Because by default sg.config.yaml sets the `CODEINTEL_PG*` and
`CODEINSIGHTS_PG*` variables to the same values as `PG*`. Later the
embedded DB logic updates the `PG*` variables but that doesn't affect
the other ones because they have been set explicitly.

There are multiple ways to fix this:
- We could explicitly clear the env vars in sg.config.yaml, so that they
  are set when initializing the embedded DB.
- We can always overwrite the codeintel and codeinsights env vars.

I chose the latter approach here because that seems more in line with
the purpose of `useSinglePostgreSQLDatabase`. At the moment it's not
obvious that the different services might use different databases even
after calling this functions.

Additional notes:

- This only affects `sg start` because it explicitly sets the codeintel
  and codeinsights env vars, causing `useSingle...` to "fail".
- It probably never worked on macOS either. Instead devs had probably a
  PostgreSQL DB running.
- With this change it's not possible anymore to point codinsights and
  codeintel to different DBs in dev mode, but I don't know it that's
  actually important.