Cody web: Fix Notepad overlap by moving the notepad button to the left when chat is visible (#50918)
Fixes #50912

I’m not proud of this one but unfortunately fixing this in a nicer way
would require a lot of refactoring between where we render the `<Notepad
/>` and `<CodyChat />`.

Right now `<Notepad />` is rendered in the root layout and absolutely
positioned to the viewport. `<CodyChat />` is rendered within the
`<RepoContainer />` and as such hierarchically _underneath_ the search
bar.

In an ideal world we can do either of the following:

- Create a shared layout for the search and repo routes and render these
components on the same layer
- Move `<CodyChat />` sidebar upwards to the root (even though this
might mean that the search size for the repo container will be
impacted). This is planned for the future (maybe) but for now, we'll
have to unblock this issue which has come up quite a lot in dogfooding
and will also come up with our customers.

The solution proposed here is, yikes, a global store to track the
current cody sidebar size and state. 🥺

## Test plan


https://user-images.githubusercontent.com/458591/233377795-f314c665-6bda-414c-bffc-790ec6d54adf.mov


<!-- All pull requests REQUIRE a test plan:
https://docs.sourcegraph.com/dev/background-information/testing_principles
-->