Fix pipeline generator
The previous code was working with a small exception: because we were
piping the entire output of `bazel run`, on a cold agent, we could get
some bazelisk output getting printed on stdout and messing with the
generated yaml.

This PR instead of building and running in a single step, does it in two
steps, which ensure that the generated yaml is solely the output of the
pipeline generator.

Test Plan: CI