Embeddings: quantize to int8 and improve perf (#50953)
This cuts memory usage of the index by ~4x by using int8 quantizations
of the float32 embeddings. Additionally, seemingly since int8s are so
much more cache efficient, this speeds up search by ~4x. Ignore the
benchmarks in the commits -- they are a bit out of date.

The reduction in precision comes at a cost in the form of less accurate
ranking/recall. However, we still get ~94% recall with these
quantizations, which seems well within the bounds of acceptable.
Notably, I chose not to use a more complex quantization scheme (like
windowing to min/max) because 1) it means scores cannot be comparable
across repos, and 2) it means we need to download the full embedding
before quantizing, which means it has to fit in memory.

On the Sourcegraph repo, compared to the float32 embeddings, int8
embeddings had a retrieval rate of 90% at Top10, 93% at Top100, and 94%
at Top1000. The mean rank of each embedding changed by only 1.2%.

There are a few somewhat independent parts to this PR, but each is
isolated to its own standalone commit. Commit messages are descriptive,
so I'd recommend reviewing commit-by-commit. The messages include
benchmark comparisons where appropriate to justify the changes.