Embeddings: fix unmarshaling old version of index (#51070)
Unmarshalling indexes 2 versions back would fail because the struct we
were unmarshalling into changed shape. This was not caught by tests
because the struct that we were using to generate the encodings was the
same as the struct we were using to unmarshal (this is the kind of thing
Protobuf's back/forward compability guarantees were intended to solve).

This PR fixes the issue by reconstructing the previous shape of the
struct and providing a conversion method to the new version. This is a
band-aid fix, and we _really_ need to solve migrations for embeddings
indexes in the near term.