Cody: Add a cache for inline completions (#51046)
This adds a simple LRU cache for completions with the main goal to
further reduce the number of requests and reduce annoying churn in the
scenario where a user receives a completion and types the exact same
characters. Previously, the new characters would cause a new completion
request which could yield different result.

Here's an example of how this cache works:

- Imagine the current file prefix looks like this with the cursor at the
end:
    ```ts
    const files = ["a", "b"];
    for(let i = 0;
    ```
- We receive the following completion:
   - ` i < files.length; i++) {`
- We now generate create different versions (up until the next `\n`) of
the input prefix by concatenating characters from the completion (Only
using the last line here to visualize):
   1. `for(let i = 0;`
   2. `for(let i = 0; `
   3. `for(let i = 0; i`
   4. `for(let i = 0; i `
   5. `for(let i = 0; i <`
   6. `for(let i = 0; i < `
   7. `for(let i = 0; i < f`
   8. `for(let i = 0; i < fi`
   9. `for(let i = 0; i < fil`
   10. `for(let i = 0; i < file`
   11. ...

## Additional thoughts

- I have't added a cache to multiline providers, since these are
triggered less often than inline suggestions anyways.
- The LRU cache is limited to 500 file prefixes, regardless of how large
these are. We might want to tweak this later. It also currently retain
the `prompt` as part of the `Completion` interface which may not be
necessary.
- I've re-enabled the request that forces adds a `\n` to the prefix.
These can now be reused if you type enter and will result in a faster
suggestion for the next line.

## Test plan

I've added a `console.log` when a cache hit is encountered to visualize
it while playing around with it:



https://user-images.githubusercontent.com/458591/234050774-2215a146-904d-47ae-b82e-c90ef131fe3e.mov


<!-- All pull requests REQUIRE a test plan:
https://docs.sourcegraph.com/dev/background-information/testing_principles
-->