rewrite as sg command, target embeddings svc
I am not yet sure if and how this will run as part of CI so I think for
now a CLI is better than a go-test.

I changed my mind about whether to target the graphql API or the
embeddings service: Targeting the grahql API requires the caller to
create token which adds uneccesary friction. At the same time we don't
(yet) alter the response in the GraphQL (we disabled ranks for now).
Let's keep it simple and change things when neeeded.